progress bar fix

+++ utils/src/main/java/observers/Progressable.java
@@ -1,5 +1,5 @@
 package observers;
 
 public interface Progressable extends Observable {
-  int getProgress();
+  float getProgress();
 }
